Module alpaca::parsing

Contains parsing routines

Functions

parse_object_kind
Parses the object's kind from its raw representation and the associated request.

parse_objects
Parses the objects contained in an HTML page.

parse_target_size
Parses the target size of an object from its HTTP request.